# openapi.yaml
components:
  schemas:
...
    CognitiveLevel:
      enum:
        - recall
        - application
        - analysis
      type: string
...
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          example: Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          nullable: true
          type: array
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          example: https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          example: 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          example: Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          example: https://example.com/errors/example
          format: uri
          type: string
      type: object
...
    GenerateDetail:
      enum:
        - brief
        - standard
        - detailed
      type: string
    GenerateDiagramType:
      enum:
        - auto
        - flowchart
        - sequence
        - class
        - state
        - er
      type: string
    GenerateKind:
      enum:
        - flashcards
        - quiz
        - mindmap
        - diagram
      type: string
    GenerateQuestionType:
      enum:
        - mcq
        - multi
        - boolean
        - short
        - open
        - ordering
        - matching
      type: string
    GenerateReq:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: https://example.com/schemas/GenerateReq.json
          format: uri
          readOnly: true
          type: string
        chapters:
          items:
            type: string
          type: array
        count:
          format: int64
          maximum: 50
          minimum: 1
          type: integer
        detail:
          $ref: "#/components/schemas/GenerateDetail"
          default: standard
        diagramType:
          $ref: "#/components/schemas/GenerateDiagramType"
          default: auto
        fileIds:
          items:
            type: string
          type: array
        format:
          type: string
        kind:
          $ref: "#/components/schemas/GenerateKind"
        length:
          type: string
        levels:
          items:
            $ref: "#/components/schemas/CognitiveLevel"
          type: array
        style:
          type: string
        title:
          maxLength: 120
          minLength: 1
          type: string
        types:
          default:
            - mcq
          items:
            $ref: "#/components/schemas/GenerateQuestionType"
          minItems: 1
          type: array
      required:
        - kind
        - count
        - title
      type: object
...
info:
  title: Capy Notebook API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /api/workspaces/{id}/generate:
    post:
      operationId: generate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReq"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Generate a quiz, flashcards, mindmap, or diagram
      tags:
        - Generate
...