# server/internal/httpapi/server.go
package httpapi

import (
	"context"
	"crypto/rand"
	"encoding/hex"
	"encoding/json"
	"errors"
	"math/big"
	"net/http"
	"strconv"
	"strings"
	"sync"
	"sync/atomic"

	"github.com/danielgtaylor/huma/v2/adapters/humachi"
	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
	"github.com/redis/go-redis/v9"
	"github.com/stripe/stripe-go/v82"

	"github.com/samyung0/capy-notebook/server/internal/auth"
	"github.com/samyung0/capy-notebook/server/internal/bank"
	"github.com/samyung0/capy-notebook/server/internal/billing"
	"github.com/samyung0/capy-notebook/server/internal/blob"
	"github.com/samyung0/capy-notebook/server/internal/mail"
	"github.com/samyung0/capy-notebook/server/internal/models"
	"github.com/samyung0/capy-notebook/server/internal/obs"
	"github.com/samyung0/capy-notebook/server/internal/pipeline"
	"github.com/samyung0/capy-notebook/server/internal/ratelimit"
	"github.com/samyung0/capy-notebook/server/internal/sourceupload"
	"github.com/samyung0/capy-notebook/server/internal/store"
)

// corsOrigins falls back to "*" when no allowlist is configured. Bearer tokens
// are sent explicitly rather than as cookies and AllowCredentials is false, so
// the wildcard is not a session-theft vector; it is still narrowed in
// production so a hostile page cannot drive the API from a user's browser.
func corsOrigins(configured []string) []string {
	if len(configured) == 0 {
		return []string{"*"}
	}
	return configured
}

// Config holds gateway settings for auth and billing. Provider OAuth
// (Google/Microsoft) is managed entirely by Clerk.
type Config struct {
	ReleaseSHA         string
	ClerkSecretKey     string
	ClerkWebhookSecret string
	AuthDisabled       bool
	DevUserID          string
	// E2EAuth enables X-E2E-User-Id identity headers (disposable E2E only).
	E2EAuth                bool
	E2ESecret              string
	E2EUserIDs             []string
	StripeSecretKey        string
	StripeWebhookSecret    string
	StripePricePro         string
	AppURL                 string
	EmailUnsubscribeSecret string
	CollaborationSecret    string
	CollaborationURL       string
	// AllowedOrigins is the CORS allowlist. Empty means "*", which is what dev
	// and e2e run with; production sets it once the SPA and API live on
	// different hostnames.
	AllowedOrigins []string
	// RateLimit governs per-user and per-IP request limits. The zero value
	// disables limiting entirely.
	RateLimit ratelimit.Config
	// ModelRegistry is the process-wide model config cache. Nil disables
	// per-model pricing and pinning (tests).
	ModelRegistry *models.Registry
	// PipelineSecret authenticates the retrieval service's and the ingest
	// host's callbacks into /api/internal/*. Empty disables those routes and
	// therefore Drive/OneDrive imports, which need the import worker.
	PipelineSecret string
	// ParseMaxPages and ParseMaxOCRPages are the ingest host parser's page cap
	// and OCR page cap, from the same env keys (CAPY_PARSE_MAX_PAGES,
	// CAPY_PARSE_OCR_PAGE_CAP). The upload policy serves them to the browser
	// and a fast-parse reservation claiming more pages is refused.
	ParseMaxPages    int
	ParseMaxOCRPages int
	// MailRecorder exposes delivered mail to Playwright. Non-nil only under
	// APP_ENV=e2e.
	MailRecorder     mail.Recorder
	MailSender       mail.Sender
	Bank             *bank.Store
	BankAssets       bank.AssetWriter
	BankAssetsURL    string
	BankCommentEmail string
}

type api struct {
	s                  *store.Store
	wh                 webhookStore
	blob               blob.Store
	pipe               *pipeline.Client
	rdb                *redis.Client
	parser             string
	cfg                Config
	mailRecorder       mail.Recorder
	limiter            *ratelimit.Limiter
	modelReg           *models.Registry
	stripeSubscription func(string) (*stripe.Subscription, error)
	stripeEntitlements func(string) ([]*stripe.Subscription, error)
	streamMu           sync.Mutex
	streamByUser       map[string]int
	streamTotal        int
	// broker fans the ingest and notification channels out from one Redis
	// subscription, and the workspace tree channel from one Postgres LISTEN
	// (treeLive says whether that LISTEN is attached); see fanout.go.
	broker   *channelBroker
	treeLive atomic.Bool
}

// New builds the full HTTP handler. huma owns every JSON operation (and the
// live OpenAPI spec at /openapi.yaml + docs at /docs); a handful of endpoints
// huma can't model — streaming SSE, blob download redirects,
// webhooks, and the pipeline chat passthrough — stay on raw chi and are
// intentionally absent from the spec. /api/internal/* stays off Huma so Orval
// does not generate a browser client for the service-to-service secret.
func New(s *store.Store, b blob.Store, pipe *pipeline.Client, rdb *redis.Client, parser string, cfg Config) http.Handler {
	billing.Init(billing.Config{SecretKey: cfg.StripeSecretKey})
	a := &api{
		s:                  s,
		wh:                 s,
		blob:               b,
		pipe:               pipe,
		rdb:                rdb,
		parser:             parser,
		cfg:                cfg,
		mailRecorder:       cfg.MailRecorder,
		limiter:            ratelimit.New(rdb, cfg.RateLimit),
		modelReg:           cfg.ModelRegistry,
		stripeSubscription: billing.RetrieveSubscription,
		stripeEntitlements: billing.ListEntitlingSubscriptions,
		streamByUser:       make(map[string]int),
	}
	if rdb != nil {
		a.broker = newChannelBroker(rdb, "ingest:*", "notif:*")
		go a.listenWorkspaceTree(context.Background())
	}
	r := chi.NewRouter()
	// Trace first so the recovery handler and every log line below it can name
	// the request; access logging second so it records panics as 500s.
	r.Use(obs.Middleware)
	r.Use(obs.AccessLog)
	r.Use(middleware.Recoverer)
	r.Use(obs.SentryMiddleware)
	r.Use(cors.Handler(cors.Options{
		AllowedOrigins: corsOrigins(cfg.AllowedOrigins),
		AllowedMethods: []string{"GET", "POST", "PATCH", "PUT", "DELETE", "OPTIONS"},
		AllowedHeaders: []string{
			"Content-Type", "Authorization", obs.HeaderTraceparent,
			auth.HeaderE2EUserID, auth.HeaderE2ESecret,
		},
		ExposedHeaders:   []string{obs.HeaderRequestID},
		AllowCredentials: false,
		MaxAge:           600,
	}))
	// Summary failures must not be cached either: visibility is read live.
	r.Use(func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			if strings.HasPrefix(r.URL.Path, "/api/public/workspaces/") {
				w.Header().Set("Cache-Control", "no-store")
			}
			next.ServeHTTP(w, r)
		})
	})
	r.Use(auth.Middleware(auth.Config{
		SecretKey:  cfg.ClerkSecretKey,
		Disabled:   cfg.AuthDisabled,
		DevUserID:  cfg.DevUserID,
		E2EAuth:    cfg.E2EAuth,
		E2ESecret:  cfg.E2ESecret,
		E2EUserIDs: cfg.E2EUserIDs,
		Store:      s,
		PublicPrefix: []string{
			"/api/email/unsubscribe",
			// Service-to-service, authenticated by X-Pipeline-Secret inside the
			// handler; there is no Clerk session to verify.
			"/api/internal/",
		},
		PublicReadPrefix: []string{
			"/api/public/workspaces/",
		},
	}))
	// After auth: limits are per user wherever there is one, and only fall back
	// to the client IP for anonymous public reads.
	r.Use(ratelimit.Middleware(a.limiter, uid))

	// Mount huma on the chi router. Doc/spec routes register at construction, so
	// this must come after all r.Use(...) calls.
	humaAPI := humachi.New(r, humaConfig())
	registerRoutes(humaAPI, a)

	// Raw (OpenAPI-excluded) routes.
	r.Get("/healthz", healthHandler(cfg.ReleaseSHA))
	r.Post("/webhooks/clerk", a.clerkWebhook)
	r.Post("/webhooks/stripe", a.stripeWebhook)
	r.Get("/api/stream", a.stream)
	r.Get("/api/email/unsubscribe", a.emailUnsubscribe)
	r.Post("/api/email/unsubscribe", a.emailUnsubscribe)
	if cfg.E2EAuth && a.mailRecorder != nil {
		r.Get("/api/e2e/emails", a.e2eEmails)
	}
	r.Post("/api/workspaces/{id}/editor-assets/uploads", a.reserveEditorAsset)
	r.Post("/api/workspaces/{id}/editor-assets/uploads/{uploadId}/complete", a.completeEditorAssetUpload)
	r.Get("/api/editor-assets/{assetId}/resolve", a.resolveEditorAsset)
	r.Post("/api/workspaces/{id}/chat/stream", a.chatStream)
	r.Post("/api/workspaces/{id}/ai/command", a.aiCommand)
	r.Post("/api/workspaces/{id}/ai/copilot", a.aiCopilot)
	if cfg.PipelineSecret != "" {
		r.Post("/api/internal/materials", a.internalCreateMaterial)
		r.Post("/api/internal/conversations/ledger", a.internalWriteConversationLedger)
		r.Post("/api/internal/source-changes/resolve", a.internalSourceAuthority)
		r.Post("/api/internal/source-changes/caption", a.internalSourceCaption)
		r.Post("/api/internal/source-refresh/publish", a.internalSourceAuthority)
		r.Get("/api/internal/agent-operations/{operationId}", a.internalGetAgentOperation)
		r.Post("/api/internal/trash", a.internalTrash)
		r.Post("/api/internal/trash/restore", a.internalRestore)
		r.Get("/api/internal/trash", a.internalListTrash)
		r.Get("/api/internal/materials/{id}/index-text", a.internalMaterialIndexText)
		r.Post("/api/internal/documents/list", a.internalListDocuments)
		r.Post("/api/internal/documents/inspect", a.internalInspectDocument)
		r.Post("/api/internal/documents/edit", a.internalEditDocument)
		r.Post("/api/internal/provider-calls", a.internalSettleProviderCall)
		r.Post("/api/internal/import/acquire", a.internalAcquireSourceImport)
		r.Post("/api/internal/import/complete", a.internalCompleteSourceImport)
		r.Post("/api/internal/import/fail", a.internalFailSourceImport)
	}
	r.Get("/api/workspaces/{id}/sources/import-content", a.getSourceImportContent)

	return r
}

/* ------------------------------------------------------------------ helpers */

func healthHandler(releaseSHA string) http.HandlerFunc {
	return func(w http.ResponseWriter, _ *http.Request) {
		if releaseSHA != "" {
			w.Header().Set("X-Capy-Release", releaseSHA)
		}
		_, _ = w.Write([]byte("ok"))
	}
}

func writeJSON(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if v != nil {
		_ = json.NewEncoder(w).Encode(v)
	}
}

func decode(r *http.Request, v any) error { return json.NewDecoder(r.Body).Decode(v) }

// errAIUnavailable is a failed handshake with the Python retrieval service.
// Chat and generate used to invent a local answer here; the client must see
// the miss so an outage does not look like a finished turn.
var errAIUnavailable = errors.New("AI service is unavailable")

// errAgentFailed is a sanitized client-facing chat failure. Provider and
// store detail stay in logs and Sentry, not in the SSE payload.
var errAgentFailed = errors.New("the chat agent hit an internal error")

// errGenerateEmpty is a 200-shaped pipeline miss: the model ran, but the
// reply could not become a material. Distinct from errAIUnavailable so a
// blank mindmap is not reported as the retrieval process being down.
var errGenerateEmpty = errors.New("the model returned no usable material")

var errScopeNoIndexedContent = errors.New("the requested scope has no indexed content")
var errContextTooLarge = errors.New("source context exceeds the selected model's input limit")
var errSourceChanged = errors.New("sources changed while this request was running; please try again")

// failFor is fail for a user-facing request: a storage refusal carries the
// numbers when the requester is the charged account.
func (a *api) failFor(w http.ResponseWriter, r *http.Request, err error) {
	if detail := quotaDetail(err, uid(r)); detail != nil {
		obs.ResponseError(w, err)
		detail["code"], detail["message"] = "storage_quota_exceeded", "storage quota exceeded"
		writeJSON(w, http.StatusForbidden, detail)
		return
	}
	a.fail(w, err)
}

func (a *api) fail(w http.ResponseWriter, err error) {
	obs.ResponseError(w, err)
	if errors.Is(err, store.ErrNotFound) {
		writeJSON(w, http.StatusNotFound, map[string]string{"message": "not found"})
		return
	}
	if errors.Is(err, store.ErrModelRefRequired) {
		writeJSON(w, http.StatusBadRequest, map[string]string{
			"code":    "model_ref_required",
			"message": "a model preference is required",
		})
		return
	}
	if errors.Is(err, store.ErrModelUnavailable) {
		writeJSON(w, http.StatusUnprocessableEntity, map[string]string{
			"code":    "model_unavailable",
			"message": "LLM model not available",
		})
		return
	}
	if errors.Is(err, store.ErrMaterialConflict) || errors.Is(err, store.ErrMaterialIDTaken) ||
		errors.Is(err, store.ErrOperationConflict) {
		writeJSON(w, http.StatusConflict, map[string]string{
			"code":    "operation_conflict",
			"message": "this operation id was already used with a different request",
		})
		return
	}
	if errors.Is(err, store.ErrTitleTaken) {
		writeJSON(w, http.StatusConflict, map[string]string{
			"code":    "title_taken",
			"message": "a material with this name already exists in this workspace",
		})
		return
	}
	if errors.Is(err, store.ErrStorageQuotaExceeded) {
		// The code alone; failFor adds the charged account's own numbers.
		writeJSON(w, http.StatusForbidden, map[string]any{
			"code":    "storage_quota_exceeded",
			"message": "storage quota exceeded",
		})
		return
	}
	var fileLimit *store.FileLimitExceededError
	if errors.As(err, &fileLimit) {
		writeJSON(w, http.StatusForbidden, map[string]any{
			"code":           fileLimit.Code(),
			"message":        "workspace file limit exceeded",
			"filesUsed":      fileLimit.Used,
			"filesReserved":  fileLimit.Reserved,
			"filesRequested": fileLimit.Requested,
			"filesLimit":     fileLimit.Limit,
			"workspaceId":    fileLimit.WorkspaceID,
		})
		return
	}
	var workspaceLimit *store.WorkspaceLimitExceededError
	if errors.As(err, &workspaceLimit) {
		writeJSON(w, http.StatusForbidden, map[string]any{
			"code":                "workspace_limit_exceeded",
			"message":             "owned workspace limit exceeded",
			"workspacesUsed":      workspaceLimit.Used,
			"workspacesRequested": workspaceLimit.Requested,
			"workspacesLimit":     workspaceLimit.Limit,
		})
		return
	}
	if errors.Is(err, store.ErrInvalidLLMKey) {
		writeJSON(w, http.StatusBadRequest, map[string]string{
			"code":    "invalid_llm_key",
			"message": "the provider rejected this key",
		})
		return
	}
	if errors.Is(err, store.ErrLLMKeyFailed) {
		writeJSON(w, http.StatusBadRequest, map[string]string{
			"code":    "llm_key_failed",
			"message": "Something went wrong, please double check if the key is valid",
		})
		return
	}
	var busy *providerBusyError
	if errors.As(err, &busy) {
		obs.ResponseError(w, obs.ExpectedError(err))
		w.Header().Set("Retry-After", strconv.Itoa(busy.retryAfter()))
		writeJSON(w, http.StatusServiceUnavailable, map[string]any{
			"code":              "provider_busy",
			"message":           busy.Error(),
			"retryAfterSeconds": busy.retryAfter(),
		})
		return
	}
	if errors.Is(err, errAIUnavailable) {
		writeJSON(w, http.StatusServiceUnavailable, map[string]string{
			"code":    "ai_unavailable",
			"message": errAIUnavailable.Error(),
		})
		return
	}
	if errors.Is(err, errGenerateEmpty) {
		writeJSON(w, http.StatusBadGateway, map[string]string{
			"code":    "generate_empty",
			"message": errGenerateEmpty.Error(),
		})
		return
	}
	if errors.Is(err, errScopeNoIndexedContent) {
		writeJSON(w, http.StatusBadRequest, map[string]string{
			"code":    "scope_has_no_indexed_content",
			"message": errScopeNoIndexedContent.Error(),
		})
		return
	}
	if errors.Is(err, errContextTooLarge) {
		writeJSON(w, http.StatusBadRequest, map[string]string{
			"code":    "context_too_large",
			"message": errContextTooLarge.Error(),
		})
		return
	}
	if errors.Is(err, errSourceChanged) {
		writeJSON(w, http.StatusConflict, map[string]string{
			"code":    "source_changed",
			"message": errSourceChanged.Error(),
		})
		return
	}
	if errors.Is(err, store.ErrTooManyLLMLeases) {
		w.Header().Set("Retry-After", "10")
		writeJSON(w, http.StatusTooManyRequests, map[string]any{
			"code":    "too_many_streams",
			"message": "too many AI requests in progress",
		})
		return
	}
	if errors.Is(err, store.ErrTooManyIngestLeases) {
		writeJSON(w, http.StatusTooManyRequests, map[string]any{
			"code":    "too_many_ingest_leases",
			"message": "too many ingest jobs in progress",
		})
		return
	}
	// Distinct from storage_quota_exceeded on purpose: this one is about the
	// caller's own inference budget, the other is about the workspace owner's
	// disk. They render as completely different messages and only one of them
	// is actionable by the person reading it.
	var credits *store.CreditsExhaustedError
	if errors.As(err, &credits) {
		writeJSON(w, http.StatusForbidden, map[string]any{
			"code":                  "llm_credits_exhausted",
			"message":               "monthly AI credits exhausted",
			"creditsUsedMicros":     credits.UsedMicros,
			"creditsReservedMicros": credits.ReservedMicros,
			"creditsLimitMicros":    credits.LimitMicros,
			"planTier":              string(credits.PlanTier),
		})
		return
	}
	var locked *store.AccountLockedError
	if errors.As(err, &locked) {
		writeJSON(w, http.StatusForbidden, map[string]any{
			"code":    locked.Code(),
			"message": "account unavailable",
			"state":   string(locked.State),
			"reason":  locked.Reason,
		})
		return
	}
	writeJSON(w, http.StatusInternalServerError, map[string]string{"message": err.Error()})
}

func id(r *http.Request) string { return chi.URLParam(r, "id") }

func uid(r *http.Request) string { return auth.UserID(r.Context()) }

// userLocale is the authenticated user's Settings locale. The pipeline uses it
// so chat/generate/editor AI replies in that language; ingest stays English.
// Unknown or missing values fall back to English. Never taken from the client.
func (a *api) userLocale(ctx context.Context, userID string) string {
	if userID == "" {
		return "en"
	}
	u, err := a.s.Me(ctx, userID)
	if err != nil {
		return "en"
	}
	return normalizeUserLocale(u.Locale)
}

func normalizeUserLocale(locale string) string {
	if locale == "zh" {
		return "zh"
	}
	return "en"
}

func randID(prefix string) string {
	b := make([]byte, 5)
	_, _ = rand.Read(b)
	return prefix + "_" + hex.EncodeToString(b)
}

func randInt(min, max int) int {
	n, _ := rand.Int(rand.Reader, big.NewInt(int64(max-min)))
	return min + int(n.Int64())
}

func (a *api) assertWS(w http.ResponseWriter, r *http.Request, wsID string) bool {
	if err := a.s.AssertWorkspaceEditor(r.Context(), uid(r), wsID); err != nil {
		if errors.Is(err, store.ErrForbidden) {
			err = store.ErrNotFound
		}
		a.fail(w, err)
		return false
	}
	return true
}

// chatAccess is what the chat stream needs to know about the actor.
type chatAccess struct {
	// canEdit follows the effective role (owner, member editor, or share-role
	// editor) unless a frozen account makes the workspace read-only: it unlocks
	// the pending-sources notice and curate. role is what the agent-tool
	// operations table is evaluated from; readOnly (frozen) and full (the
	// owner at its storage limit) narrow it.
	canEdit  bool
	readOnly bool
	full     bool
	role     store.WorkspaceRole
}

// assertWSChat admits any effective role (owner, member, or link/public
// visitor) and reports what the actor may do beyond chatting. Chat stays open
// to frozen accounts; a frozen actor or owner makes the turn read-only.
func (a *api) assertWSChat(w http.ResponseWriter, r *http.Request, wsID string) (chatAccess, bool) {
	effective, err := a.s.WorkspaceEffectiveRole(r.Context(), uid(r), wsID)
	if err != nil {
		a.fail(w, err)
		return chatAccess{}, false
	}
	readOnly, full := false, false
	if store.RoleCanEdit(effective) {
		ownerID, err := a.s.WorkspaceOwnerID(r.Context(), wsID)
		if err != nil {
			a.fail(w, err)
			return chatAccess{}, false
		}
		if readOnly, err = a.readOnly(r.Context(), ownerID); err != nil {
			a.fail(w, err)
			return chatAccess{}, false
		}
		if full, err = a.ownerFull(r.Context(), ownerID); err != nil {
			a.fail(w, err)
			return chatAccess{}, false
		}
	}
	return chatAccess{
		canEdit: store.RoleCanEdit(effective) && !readOnly, readOnly: readOnly, full: full, role: effective,
	}, true
}

func (a *api) assertWSRead(w http.ResponseWriter, r *http.Request, wsID string) bool {
	if _, err := a.s.WorkspaceAccess(r.Context(), uid(r), wsID); err != nil {
		a.fail(w, err)
		return false
	}
	return true
}

/* ------------------------------------------------------ raw source handlers */

func kindFromName(name string) string {
	return sourceupload.KindFromName(name)
}